+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/create.py
@@ -92,9 +92,9 @@ def display_path_tree(path_tree):
 @click.pass_context
 def create(ctx, name, integration_type, location, non_interactive, quiet, dry_run):
     """
-        Create scaffolding for a new integration.
+    Create scaffolding for a new integration.
 
-        NAME: The display name of the integration that will appear in documentation.
+    NAME: The display name of the integration that will appear in documentation.
     """
 
     if name.islower():