+++ datadog_checks_dev/datadog_checks/dev/tooling/github.py
@@ -3,6 +3,7 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import os
 import re
+import time
 from typing import Optional
 
 import requests
@@ -27,6 +28,29 @@ def get_auth_info(config=None):
     token = gh_config.get('token') or os.getenv('DD_GITHUB_TOKEN')
     if user and token:
         return user, token
+
+
+def get_commit(repo, commit_sha, config):
+    response = requests.get(
+        f'https://api.github.com/repos/DataDog/{repo}/git/commits/{commit_sha}', auth=get_auth_info(config)
+    )
+
+    response.raise_for_status()
+    return response.json()
+
+
+def get_tag(repo, ref, config):
+    response = requests.get(f'https://api.github.com/repos/DataDog/{repo}/git/tags/{ref}', auth=get_auth_info(config))
+
+    response.raise_for_status()
+    return response.json()
+
+
+def get_tags(repo, config):
+    response = requests.get(f'https://api.github.com/repos/DataDog/{repo}/git/refs/tags', auth=get_auth_info(config))
+
+    response.raise_for_status()
+    return response.json()
 
 
 def get_pr_labels(pr_payload):
@@ -56,12 +80,17 @@ def get_changelog_types(pr_payload):
     return changelog_labels
 
 
-def get_pr(pr_num, config=None, raw=False, org='DataDog'):
-    """
-    Get the payload for the given PR number. Let exceptions bubble up.
-    """
-    repo = basepath(get_root())
+def get_compare(base_commit, head_commit, repo, config):
+    response = requests.get(
+        f'https://api.github.com/repos/DataDog/{repo}/compare/{base_commit}...{head_commit}',
+        auth=get_auth_info(config),
+    )
 
+    response.raise_for_status()
+    return response.json()
+
+
+def get_pr_of_repo(pr_num, repo, config=None, raw=False, org='DataDog'):
     response = requests.get(PR_ENDPOINT.format(org, repo, pr_num), auth=get_auth_info(config))
 
     if raw:
@@ -71,9 +100,17 @@ def get_pr(pr_num, config=None, raw=False, org='DataDog'):
         return response.json()
 
 
+def get_pr(pr_num, config=None, raw=False, org='DataDog'):
+    """
+    Get the payload for the given PR number. Let exceptions bubble up.
+    """
+    repo = basepath(get_root())
+    return get_pr_of_repo(pr_num, repo, config, raw, org)
+
+
 def get_pr_from_hash(commit_hash, repo, config=None, raw=False):
     response = requests.get(
-        f'https://api.github.com/search/issues?q=sha:{commit_hash}+repo:DataDog/{repo}', auth=get_auth_info(config),
+        f'https://api.github.com/search/issues?q=sha:{commit_hash}+repo:DataDog/{repo}', auth=get_auth_info(config)
     )
 
     if raw:
@@ -117,3 +154,44 @@ def parse_pr_numbers(git_log_lines):
         if pr_number:
             prs.append(pr_number)
     return prs
+
+
+class Github:
+    def __init__(self, config, retry_count: int, repo: str, org: str):
+        self.__auth = get_auth_info(config)
+        self.__org = org
+        self.__repo = repo
+        self.__retry_count = retry_count
+
+    def get_user(self, login):
+        return self.__get_request_retry(f'{API_URL}/users/{login}')
+
+    def get_team_members(self, team):
+        return self.__get_request_retry(f'{API_URL}/orgs/{self.__org}/teams/{team}/members')
+
+    def get_reviews(self, pr_num):
+        return self.__get_request_retry(f'{API_URL}/repos/{self.__org}/{self.__repo}/pulls/{pr_num}/reviews')
+
+    def get_last_prs(self, user):
+        return self.__get_request_retry(
+            f'{API_URL}/search/issues?q=repo:{self.__org}/{self.__repo}+author:{user}+is:pr+sort:created'
+        )
+
+    def __get_request_retry(self, url):
+        wait = 3
+        for _ in range(self.__retry_count):
+            try:
+                return self.__get_request(url)
+            except requests.exceptions.HTTPError as e:
+                if e.response.status_code == 403:
+                    time.sleep(wait)
+                    wait *= 2
+                else:
+                    raise e
+        raise Exception(f'Error: {url}')
+
+    def __get_request(self, url):
+        response = requests.get(url, auth=self.__auth)
+
+        response.raise_for_status()
+        return response.json()