+++ datadog_checks_dev/datadog_checks/dev/tooling/templates/integration/jmx/{check_name}/tests/conftest.py
@@ -1,0 +1,6 @@
+{license_header}import pytest
+
+
+@pytest.fixture(scope='session')
+def dd_environment():
+    yield