+++ snmp/tests/metrics.py
@@ -3,7 +3,7 @@
 # Licensed under Simplified BSD License (see LICENSE)
 
 # Generic TCP metrics for routers
-# _generic-router-tcp.yaml
+# _generic-tcp.yaml
 TCP_COUNTS = [
     'tcpActiveOpens',
     'tcpPassiveOpens',
@@ -18,11 +18,11 @@ TCP_COUNTS = [
 TCP_GAUGES = ['tcpCurrEstab']
 
 # Generic UDP metrics for routers
-# _generic-router-udp.yaml
+# _generic-udp.yaml
 UDP_COUNTS = ['udpHCInDatagrams', 'udpNoPorts', 'udpInErrors', 'udpHCOutDatagrams']
 
 # Generic network interfaces metrics for routers.
-# _generic-router-if.yaml
+# _generic-if.yaml
 IF_COUNTS = [
     'ifInErrors',
     'ifInDiscards',
@@ -44,7 +44,7 @@ IF_RATES = [
 IF_GAUGES = ['ifAdminStatus', 'ifOperStatus', 'ifSpeed', 'ifHighSpeed']
 
 # Generic IP metrics for routers
-# _generic-router-ip.yaml
+# _generic-ip.yaml
 IP_COUNTS = [
     'ipSystemStatsHCInReceives',
     'ipSystemStatsInHdrErrors',
@@ -295,7 +295,7 @@ LTM_POOL_MEMBER_RATES = [
 
 
 # Base profile metrics from BPG profile
-# _generic-router-bgp4.yaml
+# _generic-bgp4.yaml
 PEER_GAUGES = [
     'bgpPeerAdminStatus',
     'bgpPeerNegotiatedVersion',