+++ win32_event_log/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - win32_event_log
 
+## 2.6.0 / 2020-08-19
+
+* [Added] Introduce new, more performant implementation. See [#7005](https://github.com/DataDog/integrations-core/pull/7005).
+
 ## 2.5.0 / 2020-06-29
 
 * [Added] Upgrade pywin32 to 228. See [#6980](https://github.com/DataDog/integrations-core/pull/6980).